BUG: mock ScenarioClock to fix DB reset test fail
`testAttemptInitialOrbitDetermination()` was creating an actual
`ScenarioClock` object, which creates a DB instance and inserts `Epoch`
rows. The test (or fixture) wasn't using the `reset_shared_db` fixture
which auto-handles clearing the DB.

This fixes the solution by instead mocking the `ScenarioClock` object,
since the DB behavior is not needed.

Fixes #64