Several changes.
First, this branch is the successor to the old feat/obs_rate branch. There's something super messed up with it, and it conflicts with the current develop branch with ray changes merged in. So, this branch has been created,
The goal is to implement the new changes from scratch for sensor revisit and downtime modeling.

Also, I'm going to improve the makefile.